comitar

+++ gerenciamento/src/main/java/DAO/daoLogin.java
@@ -96,7 +96,7 @@ public ModelUsuarios consultaUsuarioLogadoId(int id) {
 		ModelUsuarios modelUsuario = new ModelUsuarios();
 
 		PreparedStatement statement = connection.prepareStatement(sql);
-		statement.setLong(1, id);
+		statement.setInt(1, id);
 		ResultSet resultado = statement.executeQuery();
 
 		while (resultado.next()) {
